LayoutTests/js/typedarray-find-expected.txt

This test checks the behavior of the TypedArray.prototype.find function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.find.length is 1
PASS Int32Array.prototype.find.name is 'find'
PASS isSameFunctionForEachTypedArrayPrototype('find') is true
PASS testPrototypeReceivesArray('find', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Single Argument Testing
PASS testPrototypeFunction('find', '(keepEven)', [12, 5, 8, 13, 44], 12) is true
PASS testPrototypeFunction('find', '(keepEven)', [11, 13, 17, 13, 22], 22) is true
PASS testPrototypeFunction('find', '(keepEven)', [11, 13, 17, 13, 11], undefined) is true

2.0 Two Argument Testing
PASS testPrototypeFunction('find', '(keepEven, thisValue)', [11, 23, 11, 1, 44], 1) is true

3.0 Array Element Changing
PASS testPrototypeFunction('find', '(keepEvenAndChange)', [11, 15, 3, 12, 44], undefined, [5, 5, 5, 5, 5]) is true

4.0 Exception Test
PASS testPrototypeFunction('find', '(isBigEnoughAndException)', [12, 15, 10, 13, 44], 12) is true
PASS testPrototypeFunction('find', '(isBigEnoughAndException)', [9, 15, 10, 13, 44], false) threw exception exception from function.

5.0 Wrong Type for Callback Test
PASS testPrototypeFunction('find', '(8)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.find callback must be a function.
PASS testPrototypeFunction('find', '("wrong")', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.find callback must be a function.
PASS testPrototypeFunction('find', '(new Object())', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.find callback must be a function.
PASS testPrototypeFunction('find', '(null)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.find callback must be a function.
PASS testPrototypeFunction('find', '()', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.find callback must be a function.

PASS successfullyParsed is true

TEST COMPLETE
